Create fractal generator
I created a fractal generator using the formula for the Mandelbrot set.
JS paramters pass data to GLSL code, within which the vertex and fragment
shaders are created. The vertex shader transforms vertices from 3D to 2D
space and the fragment shader implements the algorithm, and colors the
points based on whether they are in the set and or how quickly they escape.

- GLSL is used for more effienct BPU-based computation.
- Coloring scheme uses HSV to RGB conversion for smooth coloring gradients
- Parameters can be adjusted by changing the function arguments
    - I will use this later to implement controls for the user