Fix blackout on zoom bug
A bug was occuring wherin when the user zoomed, the fractal would go to
black. I fixed this by using the three.js orthographic camera module. In
this way I ensured that the fractal maintained its shape and proportions
as you zoom in or out, which I think may have been the issue.